% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{download}
\usage{
download(location, version = 3, format = "excel", quiet = FALSE)
}
\arguments{
\item{location}{A character string specifying path of the location where the downloaded database should be saved}

\item{version}{Version Number of the database. Defaults to most recent version 3.0}

\item{format}{A character string specifying the format of zipped database : excel or access. Defaults to excel version of the database. The format argument is
Case Sensitive.}

\item{quiet}{If TRUE, suppress status messages (if any), and the progress bar. Defaults to FALSE}
}
\value{
None
}
\description{
Downloads the zipped RAM Legacy Stock Assessment Data Base to a specified location
}
\examples{
download(location = '../data/version2.5', version = 2.5)

download(location = '../data/version3.0', version = 3.0)

# if version is not provided then it defaults to most recent (3.0) version
download(location = '../data/version3.0')
}
