% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ramlegacy.R
\name{download_ramlegacy}
\alias{download_ramlegacy}
\title{download_ramlegacy}
\usage{
download_ramlegacy(version = NULL)
}
\arguments{
\item{version}{Version Number of the database. So far available versions are 1.0, 2.0, 2.5, 3.0 and 4.3. If version argument is not specified then it defaults to most recent version 4.3.}
}
\description{
Downloads the excel version of RAM Legacy Stock Assesment Database as R Binary object to a local directory whose path
is OS specific and chosen by \code{rappdirs} package. This path can be viewed by calling the function \code{ram_dir}.
}
