% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ramlegacy.R
\name{load_ramlegacy}
\alias{load_ramlegacy}
\title{load_ramlegacy}
\usage{
load_ramlegacy(version = NULL)
}
\arguments{
\item{version}{version number of the database that should be loaded}
}
\value{
None
}
\description{
load_ramlegacy
}
