% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ramlegacy.R
\name{read_ramlegacy}
\alias{read_ramlegacy}
\title{read_ramlegacy}
\usage{
read_ramlegacy(path, sheet = NULL, read_all = FALSE)
}
\arguments{
\item{path}{Path to the zipped version of the excel database}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). If not specified then defaults to the first
sheet.}

\item{read_all}{If TRUE, then the function will read in all the dataframes from the database}
}
\value{
A [tibble][tibble::tibble-package].If read_all is TRUE, then returns a list of tibbles from which user
can obtain individual tibbles
}
\description{
Reads different sheets of the excel database as tibbles
}
