% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ramlegacy.R
\name{download_ramlegacy}
\alias{download_ramlegacy}
\title{download_ramlegacy}
\usage{
download_ramlegacy(path, version = 3, format = "excel",
  quiet = FALSE, offline = FALSE)
}
\arguments{
\item{path}{A character string specifying path of the location where the downloaded database should be saved}

\item{version}{Version Number of the database. So far available versions : (1.0, 2.0, 2.5, 3.0). If version is not specified then it defaults to most recent version 3.0.}

\item{format}{A character string specifying the format of zipped database : excel or access. Defaults to excel version of the database. The format argument is
Case Sensitive.}

\item{quiet}{If TRUE, suppress status messages (if any), and the progress bar. Defaults to FALSE}

\item{offline}{If website is offline then downloads the database from github}
}
\value{
None
}
\description{
Downloads the zipped RAM Legacy Stock Assessment Data Base to a specified location
}
\examples{
download_ramlegacy(path = tempdir(), version = 2.5, format = 'excel')

download_ramlegacy(path = tempdir(), version = 3.0)

# if version is not provided then it defaults to most recent (3.0) version
download_ramlegacy(path = tempdir())
}
