% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ramlegacy.R
\name{load_ramlegacy}
\alias{load_ramlegacy}
\title{Read-in downloaded RAM Legacy Database}
\usage{
load_ramlegacy(version = NULL, tables = NULL, ram_path = NULL)
}
\arguments{
\item{version}{A character vector of length 1 specifying the version number of the
database. As of February 2019, the available versions are "1.0", "2.0", "2.5", "3.0", "4.3",
"4.40", "4.41" and "4.44". If version argument is not specified then it defaults to latest version
(currently "4.44"). If you want to load multiple versions please
load them one at a time as passing them all at once will throw an error.}

\item{tables}{A character vector specifying the names of particular dataframes to load from a
certain version. If not specified then a list containing all the dataframes within the requested version
is returned.}

\item{ram_path}{path to the local directory where the specified version of
the RAM Legacy Stock Excel Assessment Database was downloaded.
By default this path is set to within the rapddirs directory and can be viewed using calling the function
\code{\link{ram_dir}} and specifying the version number inside.
Although this is not the \strong{recommended} approach \code{load_ramlegacy} supports reading in the
 database's dataframes from a user-specified path provided that the database is present
 at the specified path as an rds object.}
}
\description{
Returns a list of specific dataframes or a list of all the dataframes present in the
 requested version of the database.
}
\examples{
\donttest{
# If version is not specified then current latest version
# returns a list containing area and bioparams tables from version 4.44 database
load_ramlegacy(version = "4.44", tables = c("area", "bioparams"))
}
}
\seealso{
Other ramlegacy functions: \code{\link{download_ramlegacy}},
  \code{\link{ram_dir}}
}
\concept{ramlegacy functions}
