% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ramlegacy.R
\name{load_ramlegacy}
\alias{load_ramlegacy}
\title{Read-in downloaded RAM Legacy Database}
\usage{
load_ramlegacy(version = NULL, path = NULL)
}
\arguments{
\item{version}{A character vector of length 1 specifying the version number of the
database. As of January 2019, the available versions are
"1.0", "2.0", "2.5", "3.0", "4.3", and "4.4". If version argument is not specified then it
defaults to latest version (currently "4.4"). If you want to load multiple versions please
load them one at a time as passing them all at once will throw an error.}

\item{path}{path to the local directory where the specified version of
the RAM Legacy Stock Excel Assessment Database was downloaded.
 This path can be viewed using calling the function \code{\link{ram_dir}} and specifying
 the version number inside the function call. This function \strong{does not} support
 setting a user-specified path so \strong{please
 do not pass} in a path argument to \code{path}.}
}
\description{
Loads all the dataframes present in the specified version of the database into
 user's global environment
}
\examples{
\dontrun{
# If version is not specified then current latest version
# (4.3) will be loaded
load_ramlegacy()

# load version 3.0
load_ramlegacy(version = "3.0")

# load version 2.5
load_ramlegacy(version = "2.5")
}

}
