% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{make_ramlegacy}
\alias{make_ramlegacy}
\title{Make a local sqlite3 database with the RAM Legacy Stock Stock Assessment
Database}
\usage{
make_ramlegacy(zip_file = "RLSADB_v4.3_(assessment_data_only)_access.zip",
  acc_file = "RLSADB_v4.3_(assessment_data_only).accdb")
}
\arguments{
\item{zip_file}{Name of the Access .zip file you'd like to download from
\url{http://ramlegacy.org}}

\item{acc_file}{Name of the Access .accdb file inside the zip file you'd like
to download from \url{http://ramlegacy.org}.}
}
\value{
A sqlite3 database named \code{ramlegacy.sqlite3} and \code{.csv}
  files named after each table in the database.
}
\description{
Downloads the latest Microsoft Access version (currently version 3.0) of the
RAM Legacy Stock Stock Assessment Database from \url{http://ramlegacy.org} and
converts the database to a local sqlite3 database named
\code{ramlegacy.sqlite3}. As a byproduct, \code{.csv} versions of each table
are left in the working directory. Note that you must have the utility
\code{mdb-tables} installed and in your path. You can find installation
instructions at \url{http://mdbtools.sourceforge.net}. If you're on OS X and
using homebrew, you can install it with \code{brew install mdbtools}.
}
\examples{
\dontrun{
make_ramlegacy()
}
}
