% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R
\name{make_ramlegacy}
\alias{make_ramlegacy}
\title{Make a local sqlite3 database with the RAM Legacy Stock Stock Assessment
Database}
\usage{
make_ramlegacy()
}
\description{
Downloads the latest Microsoft Access version of the RAM Legacy Stock Stock
Assessment Database from http://ramlegacy.org and converts the database to
a local sqlite3 database named \code{ramlegacy.sqlite3}. As a byproduct,
\code{.csv} versions of each table are left in the working directory.
}
\examples{
\dontrun{
make_ramlegacy()
}
}

