% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ram_dir}
\alias{ram_dir}
\title{Output OS-independent path to the zipped ramlegacy database}
\usage{
ram_dir(vers = NULL)
}
\arguments{
\item{...}{arguments potentially passed to \code{rappdirs::user_data_dir}}
}
\description{
Provides the download location for \link{download_ramlegacy} in an OS independent manner.
}
\examples{
\dontrun{
ram_dir()
}

}
