% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ram_dir}
\alias{ram_dir}
\title{Output OS-independent path to the downloaded RAM Legacy database}
\usage{
ram_dir(vers = NULL)
}
\arguments{
\item{vers}{character, version number of the database. As of January 2019,
the available versions are "1.0", "2.0", "2.5", "3.0", "4.3", and "4.4" . If version
is not specified the \code{ram_dir()} returns the path to the rappdirs directory.}
}
\description{
Provides the download location for \code{\link{download_ramlegacy}}
 in an  OS independent manner. This is also the location from where
 \code{\link{load_ramlegacy}} loads the database from.
}
\examples{
# return the path to the rappdirs directory.
ram_dir()

# Returns the path to version 4.3 subdirectory of the rappdirs directory.
# This is the path where version 4.3 of the database is downloaded to and
# read from.
ram_dir(vers = "4.3")
}
