% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramlegacy-package.R
\docType{package}
\name{ramlegacy-package}
\alias{ramlegacy}
\alias{ramlegacy-package}
\title{ramlegacy: Download and Read RAM Legacy Stock Assessment DataBase}
\description{
Contains functions to download, cache and read in
    the RAM Legacy Stock Assessment Data Base, an online compilation of
    stock assessment results for commercially exploited marine populations
    from around the world. More information about the database can be
    found at <http://ramlegacy.org/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kshtzgupta1/ramlegacy}
  \item Report bugs at \url{https://github.com/kshtzgupta1/ramlegacy/issues}
}

}
\author{
\strong{Maintainer}: Kshitiz Gupta \email{kshtzgupta1@berkeley.edu} [copyright holder]

Authors:
\itemize{
  \item Carl Boettiger \email{cboettig@gmail.com} (http://orcid.org/0000-0002-1642-628X) [copyright holder]
}

}
\keyword{internal}
