% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramlegacy-package.R
\docType{package}
\name{ramlegacy-package}
\alias{ramlegacy}
\alias{ramlegacy-package}
\title{ramlegacy: download, cache and read RAM Legacy Stock Assessment Database}
\description{
ramlegacy contains functions to download, cache and read in specified tables
from the excel version of the RAM Legacy Stock Assessment Database,
an online compilation of stock assessment results for commercially exploited
marine populations from around the world. More information about the database
can be found at <http://ramlegacy.org/>.
}
\references{
Ricard, D., Minto, C., Jensen, O.P. and Baum, J.K. (2012)
Evaluating the knowledge base and status of commercially exploited marine
species with the RAM Legacy Stock Assessment Database.
Fish and Fisheries 13 (4) 380-398. <doi:10.1111/j.1467-2979.2011.00435.x.>
}
\seealso{
\url{www.ramlegacy.org}

\url{www.github.com/ropensci/ramlegacy}

\url{www.github.com/ropensci/ramlegacy/issues}
}
\author{
\strong{Maintainer}: Kshitiz Gupta \email{kshtzgupta1@berkeley.edu} [copyright holder]

Authors:
\itemize{
  \item Carl Boettiger \email{cboettig@gmail.com} (http://orcid.org/0000-0002-1642-628X) [copyright holder]
}

Other contributors:
\itemize{
  \item Sam Albers \email{sam.albers@gmail.com} (https://orcid.org/0000-0002-9270-7884) [reviewer]
  \item Jamie Afflerbach \email{afflerbach@nceas.ucsb.edu} (https://orcid.org/0000-0002-5215-9342) [reviewer]
  \item RAM Legacy Stock Assessment Database [data contributor]
}

}
\keyword{internal}
